feat: Add `localePrefix` for navigation APIs for an improved initial render of `Link` when using `localePrefix: never`. Also fix edge case in middleware when using localized pathnames for redirects that remove a locale prefix (fixes an infinite loop). (#678)
By accepting an optional `localePrefix` for the navigation APIs, we can
get the initial render of the `href` of `Link` right if `localePrefix: 'never'` is set. This can be helpful if domain-based routing is used and
you have a single locale per domain.

Note that this change is backward-compatible. It's now recommended to
set the `localePrefix` for the navigation APIs to get improved behavior
for `Link` in case `localePrefix: 'never'` is used, but otherwise your
app will keep working with the previous behavior.





Ref #444