fix: Allow to call `getRequestConfig` in outer module closure in a Client Component module graph (#687)
Fixes #685 

This is fixed for compatibility, since we've silently supported this
pattern previously. However, it's still not recommended to call
`getRequestConfig` in a Client Component module graph as the function
can't do anything useful there. Instead, you can extract shared config
like `locales` to a file like
[`src/navigation.ts`](https://next-intl-docs.vercel.app/docs/routing/navigation)
or any other module that is then imported into Client and Server
Components.