fix: When using domain-based routing, use `defaultLocale` of a domain instead of the top-level one in case no other locale matches better on the domain (#1000)
Fixes #998 

Note that the `defaultLocale` of a domain is used if no other locale
matches better. However, if a domain supports multiple locales, the
best-matching one will be selected based on the `accept-language`
header. If you want to always use the `defaultLocale` in case no prefix
is provided, then you can turn off `localeDetection`.