docs: Remove community links (#1398)
It's unfortunately no longer possible to keep up with all the cool
content that gets published about `next-intl` and I don't want to
prioritize content selectively. Due to this, I'm removing this section
from the blog page.

A huge thanks to everyone who takes the time to publish content that
mentions `next-intl`! ❤️