fix: Handle `localeCookie` setting correctly on client side (#1417)
**Changes:**
- Move the `localeCookie` from the second argument of `createMiddleware`
to the first one and expect it to be shared with the navigation APIs
(preferably defined with `defineRouting`). By sharing the `localeCookie`
setting with the navigation APIs, this fixes a bug to handle
`localeCookie` correctly on the client side.
- Deprecate second argument of middleware entirely and move all routing
config to `defineRouting`.
- Allow to configure `name` of cookie.
- Don't allow to configure `expires` (introduces unnecessary complexity
and is considered legacy anyway—use `maxAge` instead).