feat!: Make package ESM-only, target modern browsers, use modern JSX transform (#1470)
**Changes**

1. `next-intl` as well as its base library `use-intl` are now ESM-only.
With this, the library is now ~7% smaller and provides full support for
tree shaking. Most tools support ESM out-of-the-box, but some might
require additional configuration (e.g. Jest might benefit from
[`next/jest`](https://nextjs.org/docs/app/building-your-application/testing/jest#manual-setup)).
The only exception is `next-intl/plugin` which is provided both as ESM
as well as CommonJS due to `next.config.js` still being popular.
2. Syntax is compiled down to the Browserslist
[`defaults`](https://github.com/browserslist/browserslist/blob/065c69b0e1e6ad679263d03e22c605830bbc52e5/index.js#L460)
query, which is a shortcut for `> 0.5%, last 2 versions, Firefox ESR,
not dead`. The `defaults` query was thoroughly designed by the
Browserslist community and is considered a reasonable default for modern
apps. If you target outdated browsers, you can use
[`transpilePackages`](https://nextjs.org/docs/app/api-reference/next-config-js/transpilePackages)
in combination with a
[`browserslist`](https://nextjs.org/docs/architecture/supported-browsers)
config in Next.js to compile `next-intl` for older browsers.
3. The minimum React version is 17 now (which supports the modern JSX
transform)