fix: Make locale cookie a session cookie (#1634)
Initially, I planned to add an expiration of 5 hours to the locale
cookie for `next-intl@4` to comply with GDPR regulations. However, this
has the implication that if the browser remains open for longer than 5
hours, the cookie can be reset in the middle of a session.

Due to this, it seems more reasonable to not set an expiration at all,
turning the cookie into a session cookie. Session cookies expiry only
when a browser is closed. On mobile, this can be even more beneficial,
as browsers are rarely closed (the browser can clear cookies though if
memory is constrained).