fix: Change imports to use `.js` to make them compatible with `eslint-plugin-import` (#1667)
@amannn Following our conversation on X, these are the changes that I've
made that appear to be fixing the imports when the `import/named` ESLint
rule is used.

As mentioned before, I also took the liberty of adding the ESLint rule
`'import/extensions': 'error'` to ensure the right extensions are used
in the import/export statements. For context, when compiling to ESM
modules, TS copies over the import/export statements as-is, so during
development, runtime compatible file extensions must be used.

For local testing I ran a build and then manually copied over the `dist`
folder into `node_modules/next-intl` in my other project.