feat: change createMessagesDeclaration to also accept array for monorepo support (#1700)
Resolves: https://github.com/amannn/next-intl/issues/1699

This change renames "createMessagesDeclaration" to
"createMessagesDeclarations" and changes the type to an array. Each
entry of the array will be used to create message declaration files.
Supporting multiple message files helps to use next-intl in monorepo
setups.

Since v4 is in beta, I renamed the property and removed support for a
single string. To me, clean types lead to cleaner code. If compatibility
is a more of a concern during the beta phase, I could not change the
property name and keep the support for single strings. Please let me
know.

---------

Co-authored-by: Jan Amann <jan@amann.work>