chore: add MDX files to tsconfig include pattern
This makes MDX files part of the TypeScript program in the editor. As a
result, MDX IntelliSense uses the same configuration as the rest of the
program. This also means TypeScript only needs to provide IntelliSense
for one program instead of using a default fallback for MDX. This
improves performance.