fix: Avoid `usePathname` inconsistency in Next.js leading to a hydration error with custom prefixes, `localePrefix: 'always'` and static rendering (#2012)
Fixes #2011

This PR introduces a workaround for
[vercel/next.js#73085](https://github.com/vercel/next.js/issues/73085).

In Next.js, using `usePathname` in combination with rewrites and
middleware can trigger hydration errors. These only occur in production
builds, and while they are documented, the information is somewhat
hidden:

<img width="480" height="672" alt="CleanShot 2025-09-08 at 19 48 12@2x"
src="https://github.com/user-attachments/assets/e656de20-c617-49ca-94e7-35000854cae5"
/>

This issue was partially addressed in
[#1571](https://github.com/amannn/next-intl/pull/1571). However, I
noticed that when `localePrefix.mode` is set to `'always'`, the
workaround was not applied. Instead of adding an extra condition for
`'always'`, I flipped the check to `localePrefix.mode !== 'never'`. As
far as I can tell this should work fine, but let me know if you see any
risks/issues with this.