docs: Update static rendering instructions for locale-based routing (#2033)
The docs suggest using `generateStaticParams` for all locales but this
can slow down builds, especially if you have lots of locales and routes.

I'm updating that section to mention it's possible to return an empty
array or use `export const dynamic = 'force-static'` and the page will
be generated on the first visit without becoming dynamic.